refactor(proposer): extract recovery orchestration
