+++ crates/proof/proposer/src/pipeline.rs
@@ -21,42 +21,43 @@
 //! work. Discard retries use retry-specific sessions because the prover service
 //! intentionally replays `Succeeded` sessions for the root-derived id.
 
-use std::{collections::HashMap, sync::Arc, time::Duration};
+#[cfg(test)]
+use std::collections::HashMap;
+use std::{sync::Arc, time::Duration};
 
-use alloy_primitives::{Address, B256};
-use async_trait::async_trait;
+use alloy_primitives::Address;
+#[cfg(test)]
+use alloy_primitives::B256;
 use base_proof_contracts::{
-    AggregateVerifierClient, AnchorStateRegistryClient, DisputeGameFactoryClient, encode_extra_data,
+    AggregateVerifierClient, AnchorStateRegistryClient, DisputeGameFactoryClient,
 };
 #[cfg(test)]
 use base_proof_primitives::ProofResult;
-use base_proof_rpc::{L1Provider, L2Provider, RollupProvider, RpcError};
+use base_proof_rpc::{L1Provider, L2Provider, RollupProvider};
 use base_prover_service_client::ProofRequesterProvider;
 use eyre::Result;
-use futures::{StreamExt, stream};
 use tokio::{sync::mpsc, task::JoinHandle};
 use tokio_util::sync::CancellationToken;
-use tracing::{debug, info, instrument, warn};
+use tracing::{info, instrument, warn};
 
 use crate::{
     Metrics,
-    driver::{DriverConfig, RecoveredState},
-    error::ProposerError,
+    driver::DriverConfig,
     output_proposer::OutputProposer,
     proof_collector::{
-        ProofCollector, ProofCollectorOrchestrator, ProofCollectorRecoveryProvider,
-        ProofCollectorRuntimeConfig, ProofCollectorState, ProofCollectorTickResult,
-        ProofRecoveryCache,
+        ProofCollector, ProofCollectorOrchestrator, ProofCollectorRuntimeConfig,
+        ProofCollectorState, ProofCollectorTickResult,
     },
     proof_dispatcher::{
         ProofDispatcher, ProofDispatcherConfig, ProofDispatcherRuntimeConfig, ProofDispatcherState,
     },
+    proof_recovery::{ProofRecovery, ProofRecoveryCache, ProofRecoveryConfig},
     proof_submitter::{ProofSubmitter, ProofSubmitterConfig},
 };
 #[cfg(test)]
 use crate::{
-    proof_collector::ProofSubmitEffect, proof_dispatcher::ProofDispatchOutcome,
-    proof_submitter::SubmitAction,
+    driver::RecoveredState, error::ProposerError, proof_collector::ProofSubmitEffect,
+    proof_dispatcher::ProofDispatchOutcome, proof_submitter::SubmitAction,
 };
 
 /// Configuration for the proving pipeline.
@@ -139,6 +140,7 @@ where
     proof_dispatcher: ProofDispatcher<L1, L2, R>,
     proof_collector: ProofCollector<R>,
     proof_submitter: ProofSubmitter<L1, R>,
+    proof_recovery: ProofRecovery<R, ASR, F>,
     l1_client: Arc<L1>,
     l2_client: Arc<L2>,
     rollup_client: Arc<R>,
@@ -164,6 +166,7 @@ where
             proof_dispatcher: self.proof_dispatcher.clone(),
             proof_collector: self.proof_collector.clone(),
             proof_submitter: self.proof_submitter.clone(),
+            proof_recovery: self.proof_recovery.clone(),
             l1_client: Arc::clone(&self.l1_client),
             l2_client: Arc::clone(&self.l2_client),
             rollup_client: Arc::clone(&self.rollup_client),
@@ -239,6 +242,19 @@ where
                 output_fetch_concurrency: config.recovery_scan_concurrency,
             },
         );
+        let proof_recovery = ProofRecovery::new(
+            ProofRecoveryConfig {
+                block_interval: config.driver.block_interval,
+                intermediate_block_interval: config.driver.intermediate_block_interval,
+                game_type: config.driver.game_type,
+                allow_non_finalized: config.driver.allow_non_finalized,
+                anchor_state_registry_address: config.driver.anchor_state_registry_address,
+                scan_concurrency: config.recovery_scan_concurrency,
+            },
+            Arc::clone(&rollup_client),
+            Arc::clone(&anchor_registry),
+            Arc::clone(&factory_client),
+        );
 
         Self {
             config,
@@ -246,6 +262,7 @@ where
             proof_dispatcher,
             proof_collector,
             proof_submitter,
+            proof_recovery,
             l1_client,
             l2_client,
             rollup_client,
@@ -265,12 +282,14 @@ where
         self.cancel = cancel;
     }
 
-    fn collector_orchestrator(&self) -> ProofCollectorOrchestrator<L1, L2, R, Self> {
+    fn collector_orchestrator(
+        &self,
+    ) -> ProofCollectorOrchestrator<L1, L2, R, ProofRecovery<R, ASR, F>> {
         ProofCollectorOrchestrator::new(
             self.proof_collector.clone(),
             self.proof_dispatcher.clone(),
             self.proof_submitter.clone(),
-            Arc::new(self.clone()),
+            Arc::new(self.proof_recovery.clone()),
             ProofCollectorRuntimeConfig {
                 block_interval: self.config.driver.block_interval,
                 max_retries: self.config.max_retries,
@@ -386,7 +405,7 @@ where
     ) {
         let _tick_timer = base_metrics::timed!(Metrics::tick_duration_seconds());
 
-        let (recovered, safe_head) = match self.try_recover_and_plan(cache).await {
+        let (recovered, safe_head) = match self.proof_recovery.try_recover_and_plan(cache).await {
             Some(pair) => pair,
             None => {
                 Metrics::pipeline_retries().set(state.retry_counts.values().sum::<u32>() as f64);
@@ -457,7 +476,7 @@ where
     ) {
         let _tick_timer = base_metrics::timed!(Metrics::collector_tick_duration_seconds());
 
-        let (recovered, safe_head) = match self.try_recover_and_plan(cache).await {
+        let (recovered, safe_head) = match self.proof_recovery.try_recover_and_plan(cache).await {
             Some(pair) => pair,
             None => {
                 Metrics::pipeline_retries().set(state.retry_counts.values().sum::<u32>() as f64);
@@ -602,379 +621,6 @@ where
         should_retry
     }
 
-    /// Attempts to recover on-chain state and fetch the safe head.
-    ///
-    /// Returns `None` if either step fails (logged as warnings), allowing the
-    /// caller to fall through to the poll-tick sleep.
-    async fn try_recover_and_plan(
-        &self,
-        cache: &mut Option<CachedRecovery>,
-    ) -> Option<(RecoveredState, u64)> {
-        if let Some(cached) = cache.as_ref() {
-            let safe_head = match self.latest_safe_block_number().await {
-                Ok(n) => n,
-                Err(e) => {
-                    warn!(error = %e, "Failed to fetch safe head, retrying next tick");
-                    return None;
-                }
-            };
-
-            let next_proposal_block =
-                match cached.state.l2_block_number.checked_add(self.config.driver.block_interval) {
-                    Some(block) => block,
-                    None => {
-                        warn!(
-                            cached_block = cached.state.l2_block_number,
-                            block_interval = self.config.driver.block_interval,
-                            "Cannot compute next proposal block, retrying next tick"
-                        );
-                        return None;
-                    }
-                };
-
-            if safe_head < next_proposal_block {
-                debug!(
-                    safe_head,
-                    cached_block = cached.state.l2_block_number,
-                    next_proposal_block,
-                    "Safe head below next proposal target, skipping recovery"
-                );
-                return Some((cached.state, safe_head));
-            }
-
-            let state = match self.recover_latest_state(cache).await {
-                Ok(s) => s,
-                Err(e) => {
-                    warn!(error = %e, "Failed to recover on-chain state, retrying next tick");
-                    return None;
-                }
-            };
-
-            return Some((state, safe_head));
-        }
-
-        let (state_result, safe_head_result) =
-            tokio::join!(self.recover_latest_state(cache), self.latest_safe_block_number(),);
-
-        let state = match state_result {
-            Ok(s) => s,
-            Err(e) => {
-                warn!(error = %e, "Failed to recover on-chain state, retrying next tick");
-                return None;
-            }
-        };
-
-        let safe_head = match safe_head_result {
-            Ok(n) => n,
-            Err(e) => {
-                warn!(error = %e, "Failed to fetch safe head, retrying next tick");
-                return None;
-            }
-        };
-
-        Some((state, safe_head))
-    }
-
-    /// Recovers the latest on-chain state using a deterministic forward walk
-    /// from the anchor root.
-    ///
-    /// # Strategy
-    ///
-    /// 1. Read `game_count` from the factory and anchor root from the registry
-    ///    once the safe head is high enough to need recovery.
-    /// 2. **Cache check — fast path.** If both `game_count` and `anchor_root`
-    ///    match the cache, return the cached state immediately (zero RPCs).
-    /// 3. **Forward walk.** Walk from the anchor block, stepping by
-    ///    `block_interval`. At each step:
-    ///    - Compute expected block number deterministically.
-    ///    - Fetch the canonical output root and intermediate roots from the
-    ///      rollup node.
-    ///    - Build `extraData` from the block number, parent address, and
-    ///      intermediate roots.
-    ///    - Call `factory.games(gameType, rootClaim, extraData)` to look up
-    ///      the game by its unique UUID.
-    ///    - If `proxy == Address::ZERO`, no game exists — gap found, stop.
-    ///    - Otherwise, advance to the returned proxy as the new parent.
-    ///
-    /// This approach is deterministic: the correct game for each step is
-    /// uniquely identified by its `(gameType, rootClaim, extraData)` tuple.
-    /// There is no ambiguity or filtering — the game either exists or it
-    /// doesn't.
-    ///
-    /// # Bounding
-    ///
-    /// The walk is NOT bounded by the safe/finalized L2 head because it
-    /// only verifies existing on-chain games (which were already submitted
-    /// and included on L1). New proposal dispatch in [`Self::dispatcher_loop`]
-    /// is separately bounded by the safe head.
-    async fn recover_latest_state(
-        &self,
-        cache: &mut Option<CachedRecovery>,
-    ) -> Result<RecoveredState, ProposerError> {
-        let count = self
-            .factory_client
-            .game_count()
-            .await
-            .map_err(|e| ProposerError::Contract(format!("recovery game_count failed: {e}")))?;
-
-        // Read the anchor root and anchor game from one L1 snapshot so
-        // recovery cannot combine an old root with a newer anchor game.
-        let anchor_snapshot = self
-            .anchor_registry
-            .anchor_snapshot()
-            .await
-            .map_err(|e| ProposerError::Contract(format!("anchor_snapshot failed: {e}")))?;
-        let anchor = anchor_snapshot.anchor_root;
-
-        // The cached tip is valid as long as the anchor hasn't advanced past
-        // it. The anchor advances when games resolve (~every 20 min after the
-        // dispute window elapses), but it always stays behind the chain tip.
-        let tip_still_valid =
-            |cached: &CachedRecovery| anchor.l2_block_number <= cached.state.l2_block_number;
-
-        // Fast path: game_count unchanged and anchor still behind tip →
-        // return the cached state with zero additional RPCs.
-        if let Some(cached) = cache.as_ref()
-            && tip_still_valid(cached)
-            && cached.game_count == count
-        {
-            debug!(game_count = count, "No changes since last recovery, returning cached state");
-            return Ok(cached.state);
-        }
-
-        // ── Forward walk ────────────────────────────────────────────────
-        //
-        // When game_count increased and the anchor is still at or behind
-        // the cached tip, resume from the tip instead of re-walking from
-        // the anchor. This turns post-submission recovery from O(K) to
-        // O(1).
-        //
-        // A full walk from the anchor is required when:
-        // - No cache exists (cold start, or invalidated by RootMismatch /
-        //   max_retries).
-        // - The anchor advanced past the cached tip (governance / anomaly).
-        // - game_count decreased (L1 reorg removed games).
-        let start = match cache.as_ref() {
-            Some(cached) if tip_still_valid(cached) && count > cached.game_count => {
-                debug!(
-                    cached_block = cached.state.l2_block_number,
-                    old_count = cached.game_count,
-                    new_count = count,
-                    "Resuming forward walk from cached tip"
-                );
-                cached.state
-            }
-            _ => {
-                let parent_address = if anchor_snapshot.anchor_game.is_zero() {
-                    self.config.driver.anchor_state_registry_address
-                } else {
-                    anchor_snapshot.anchor_game
-                };
-
-                RecoveredState {
-                    parent_address,
-                    output_root: anchor.root,
-                    l2_block_number: anchor.l2_block_number,
-                }
-            }
-        };
-
-        let state = self.forward_walk(&start).await?;
-
-        *cache = Some(CachedRecovery { game_count: count, state });
-        Ok(state)
-    }
-
-    /// Performs a deterministic forward walk to find the latest verified game
-    /// using UUID-based `games()` lookups.
-    ///
-    /// The walk starts from `start`, which is either the anchor state (full
-    /// walk) or the cached tip from a previous walk (incremental).
-    ///
-    /// At each step:
-    /// 1. Compute the expected block number: `parent_block + block_interval`.
-    /// 2. Fetch all intermediate roots (including the target block's output
-    ///    root) from the rollup node in a single batch.
-    /// 3. Build `extraData` from the block number, parent address, and
-    ///    intermediate roots.
-    /// 4. Call `factory.games(gameType, rootClaim, extraData)` — the factory
-    ///    returns the proxy address if a game with this exact UUID exists, or
-    ///    `Address::ZERO` if not.
-    /// 5. `Address::ZERO` → gap found, stop. Otherwise advance the parent.
-    ///
-    /// Because the game's UUID is computed from canonical data, there is no
-    /// ambiguity: the correct game either exists or it doesn't. Invalid games
-    /// (wrong root claim, wrong parent, wrong intermediate roots) simply have
-    /// different UUIDs and are never matched.
-    ///
-    /// The walk is sequential (each step needs the previous proxy address for
-    /// `extraData`), but each step requires only two RPCs: one
-    /// `fetch_canonical_roots` batch and one `games()` lookup.
-    async fn forward_walk(&self, start: &RecoveredState) -> Result<RecoveredState, ProposerError> {
-        let block_interval = self.config.driver.block_interval;
-        let game_type = self.config.driver.game_type;
-
-        let log_interval = (block_interval / 5).max(1);
-
-        let mut parent_address = start.parent_address;
-        let mut parent_output_root = start.output_root;
-        let mut parent_block = start.l2_block_number;
-        let mut steps: u64 = 0;
-
-        while let Some(expected_block) = parent_block.checked_add(block_interval) {
-            // Fetch all intermediate roots (including the final root at
-            // `expected_block`) from the rollup node in one batch. The last
-            // element of `intermediate_blocks` is always `expected_block`,
-            // so this also provides the canonical output root — no separate
-            // `output_at_block` call needed.
-            let intermediate_blocks =
-                self.proof_submitter.intermediate_block_numbers(parent_block)?;
-            let intermediate_roots =
-                match self.fetch_canonical_roots(intermediate_blocks.clone()).await {
-                    Ok(roots) => roots,
-                    Err(ProposerError::Rpc(RpcError::BlockNotFound(_))) => {
-                        // The block doesn't exist yet (ahead of safe head).
-                        // This is the natural termination point of the walk.
-                        debug!(
-                            block = expected_block,
-                            "Block not available yet, treating as end of walk"
-                        );
-                        break;
-                    }
-                    Err(e) => {
-                        // All other RPC errors (retryable or not) propagate so
-                        // recovery retries on the next tick rather than caching
-                        // a partial result.
-                        warn!(
-                            expected_block,
-                            parent_block,
-                            error = %e,
-                            "Forward walk failed to fetch canonical roots"
-                        );
-                        return Err(e);
-                    }
-                };
-
-            // Extract the canonical root for the target block (always the
-            // last intermediate block).
-            let canonical_root = *intermediate_roots.get(&expected_block).ok_or_else(|| {
-                ProposerError::Internal(format!(
-                    "missing canonical root for expected block {expected_block}"
-                ))
-            })?;
-
-            // Build the ordered intermediate root vector matching extraData layout.
-            let intermediate_root_vec: Vec<B256> = intermediate_blocks
-                .iter()
-                .map(|ib| {
-                    intermediate_roots.get(ib).copied().ok_or_else(|| {
-                        ProposerError::Internal(format!(
-                            "missing canonical root for intermediate block {ib}"
-                        ))
-                    })
-                })
-                .collect::<Result<Vec<_>, _>>()?;
-
-            // Build extraData and look up the game by UUID.
-            let extra_data =
-                encode_extra_data(expected_block, parent_address, &intermediate_root_vec);
-
-            let lookup =
-                self.factory_client.games(game_type, canonical_root, extra_data).await.map_err(
-                    |e| {
-                        ProposerError::Contract(format!(
-                            "games lookup failed at block {expected_block}: {e}"
-                        ))
-                    },
-                )?;
-
-            if lookup == Address::ZERO {
-                info!(
-                    gap_block = expected_block,
-                    parent_block,
-                    parent_address = %parent_address,
-                    games_verified = steps,
-                    "No game found at expected block, will propose from here"
-                );
-                break;
-            }
-
-            parent_address = lookup;
-            parent_output_root = canonical_root;
-            parent_block = expected_block;
-            steps += 1;
-
-            if steps.is_multiple_of(log_interval) {
-                info!(
-                    games_verified = steps,
-                    latest_block = parent_block,
-                    "Recovery forward walk in progress"
-                );
-            }
-        }
-
-        if steps > 0 {
-            info!(
-                latest_block = parent_block,
-                parent_address = %parent_address,
-                games_verified = steps,
-                "Recovery forward walk complete"
-            );
-        }
-
-        Ok(RecoveredState {
-            parent_address,
-            output_root: parent_output_root,
-            l2_block_number: parent_block,
-        })
-    }
-
-    /// Returns the latest safe L2 block number.
-    async fn latest_safe_block_number(&self) -> Result<u64, ProposerError> {
-        let sync_status = self.rollup_client.sync_status().await?;
-        if self.config.driver.allow_non_finalized {
-            Ok(sync_status.safe_l2.number)
-        } else {
-            Ok(sync_status.finalized_l2.number)
-        }
-    }
-
-    /// Concurrently fetches canonical output roots for the given block numbers.
-    async fn fetch_canonical_roots(
-        &self,
-        blocks: Vec<u64>,
-    ) -> Result<HashMap<u64, B256>, ProposerError> {
-        self.fetch_canonical_root_results(blocks)
-            .await
-            .into_iter()
-            .map(|(block_number, result)| result.map(|root| (block_number, root)))
-            .collect()
-    }
-
-    async fn fetch_canonical_root_results(
-        &self,
-        blocks: Vec<u64>,
-    ) -> HashMap<u64, Result<B256, ProposerError>> {
-        if blocks.is_empty() {
-            return HashMap::new();
-        }
-        stream::iter(blocks)
-            .map(|block_number| {
-                let rollup = &self.rollup_client;
-                async move {
-                    let result = rollup
-                        .output_at_block(block_number)
-                        .await
-                        .map(|out| out.output_root)
-                        .map_err(ProposerError::Rpc);
-                    (block_number, result)
-                }
-            })
-            .buffered(self.config.recovery_scan_concurrency)
-            .collect()
-            .await
-    }
-
     /// Validates the proof and submits it to L1 by delegating to the
     /// [`ProofSubmitter`].
     ///
@@ -990,23 +636,6 @@ where
         parent_address: Address,
     ) -> Result<(), SubmitAction> {
         self.proof_submitter.submit(proof_result, target_block, parent_address).await
-    }
-}
-
-#[async_trait]
-impl<L1, L2, R, ASR, F> ProofCollectorRecoveryProvider for ProvingPipeline<L1, L2, R, ASR, F>
-where
-    L1: L1Provider + 'static,
-    L2: L2Provider + 'static,
-    R: RollupProvider + 'static,
-    ASR: AnchorStateRegistryClient + 'static,
-    F: DisputeGameFactoryClient + 'static,
-{
-    async fn recover_latest_state(
-        &self,
-        cache: &mut Option<ProofRecoveryCache>,
-    ) -> Result<RecoveredState, ProposerError> {
-        Self::recover_latest_state(self, cache).await
     }
 }
 
@@ -1075,6 +704,7 @@ mod tests {
     use alloy_primitives::{Address, B256};
     use async_trait::async_trait;
     use base_proof_primitives::{ProofResult, Proposal};
+    use base_proof_rpc::RpcError;
     use base_prover_service_client::ProverServiceClientError;
     use base_prover_service_protocol::{
         GetProofRequest, GetProofResponse, ListProofsRequest, ListProofsResponse, ProofStatus,
@@ -1125,9 +755,6 @@ mod tests {
     /// Game type used across recovery tests.
     const TEST_GAME_TYPE: u32 = 42;
 
-    /// Default block interval for recovery tests (matches `DriverConfig` default).
-    const TEST_BLOCK_INTERVAL: u64 = 512;
-
     /// Default anchor block number.
     const TEST_ANCHOR_BLOCK: u64 = 0;
 
@@ -1146,63 +773,6 @@ mod tests {
         Address::new(bytes)
     }
 
-    /// Builds a chain of `N` sequential games starting from the anchor,
-    /// registering them in the factory's `uuid_games` map.
-    ///
-    /// Uses `block_interval == intermediate_block_interval == TEST_BLOCK_INTERVAL`
-    /// (one intermediate root per game, equal to the root claim).
-    ///
-    /// Returns `(factory, output_roots)` ready to use in pipeline builders.
-    fn game_chain(n: usize) -> (MockDisputeGameFactory, HashMap<u64, B256>) {
-        game_chain_full(n, TEST_ANCHOR_BLOCK, TEST_BLOCK_INTERVAL, TEST_BLOCK_INTERVAL)
-    }
-
-    /// Builds a chain of `N` sequential games with configurable intervals.
-    fn game_chain_full(
-        n: usize,
-        anchor_block: u64,
-        block_interval: u64,
-        intermediate_block_interval: u64,
-    ) -> (MockDisputeGameFactory, HashMap<u64, B256>) {
-        let mut uuid_games = std::collections::HashMap::new();
-        let mut output_roots = HashMap::new();
-        let intermediate_count = block_interval / intermediate_block_interval;
-
-        let mut parent = Address::ZERO; // anchor_state_registry_address default
-        for i in 0..n {
-            let block = anchor_block + block_interval * (i as u64 + 1);
-            let root_claim = B256::repeat_byte((i as u8) + 1);
-
-            // Build intermediate roots (canonical values).
-            let parent_block = block - block_interval;
-            let mut intermediate_roots = Vec::with_capacity(intermediate_count as usize);
-            for j in 1..=intermediate_count {
-                let ib = parent_block + j * intermediate_block_interval;
-                let ir = if ib == block { root_claim } else { B256::repeat_byte(ib as u8) };
-                output_roots.insert(ib, ir);
-                intermediate_roots.push(ir);
-            }
-            output_roots.insert(block, root_claim);
-
-            let extra_data = encode_extra_data(block, parent, &intermediate_roots);
-            let proxy = proxy_addr(i as u64);
-
-            uuid_games.insert((TEST_GAME_TYPE, root_claim, extra_data), proxy);
-
-            parent = proxy;
-        }
-
-        let factory = MockDisputeGameFactory {
-            games: Vec::new(),
-            game_count_override: Some(n as u64),
-            uuid_games,
-            games_should_fail: false,
-            game_count_calls: None,
-        };
-
-        (factory, output_roots)
-    }
-
     // ---- Pipeline builders ----
 
     /// Type alias to reduce repetition in builder return types.
@@ -1213,38 +783,6 @@ mod tests {
         MockAnchorStateRegistry,
         MockDisputeGameFactory,
     >;
-
-    #[derive(Debug)]
-    struct SnapshotOnlyAnchorStateRegistry {
-        snapshot: base_proof_contracts::AnchorSnapshot,
-    }
-
-    #[async_trait::async_trait]
-    impl AnchorStateRegistryClient for SnapshotOnlyAnchorStateRegistry {
-        async fn anchor_snapshot(
-            &self,
-        ) -> std::result::Result<
-            base_proof_contracts::AnchorSnapshot,
-            base_proof_contracts::ContractError,
-        > {
-            Ok(self.snapshot)
-        }
-    }
-
-    /// Builds a recovery pipeline with a pre-configured factory and canonical
-    /// output roots. Uses default anchor block and block interval.
-    fn recovery_pipeline(
-        factory: MockDisputeGameFactory,
-        output_roots: HashMap<u64, B256>,
-    ) -> TestPipeline {
-        recovery_pipeline_full(
-            factory,
-            output_roots,
-            TEST_ANCHOR_BLOCK,
-            TEST_BLOCK_INTERVAL,
-            TEST_BLOCK_INTERVAL,
-        )
-    }
 
     fn recovery_pipeline_full(
         factory: MockDisputeGameFactory,
@@ -1279,25 +817,6 @@ mod tests {
             block_interval,
             intermediate_block_interval,
             output_proposer,
-        )
-    }
-
-    fn recovery_pipeline_full_with_anchor_game(
-        factory: MockDisputeGameFactory,
-        output_roots: HashMap<u64, B256>,
-        anchor_block: u64,
-        anchor_game: Address,
-        block_interval: u64,
-        intermediate_block_interval: u64,
-    ) -> TestPipeline {
-        recovery_pipeline_full_with_anchor_game_and_output_proposer(
-            factory,
-            output_roots,
-            anchor_block,
-            anchor_game,
-            block_interval,
-            intermediate_block_interval,
-            Arc::new(MockOutputProposer),
         )
     }
 
@@ -1346,426 +865,6 @@ mod tests {
             output_proposer,
             cancel,
         )
-    }
-
-    // ---- Pipeline lifecycle tests ----
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_returns_anchor_when_no_games() {
-        let factory = MockDisputeGameFactory::with_games(vec![]);
-        let pipeline = recovery_pipeline(factory, HashMap::new());
-
-        let mut cache: Option<CachedRecovery> = None;
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        assert_eq!(
-            state.parent_address,
-            Address::ZERO,
-            "should return anchor_state_registry_address"
-        );
-        assert_eq!(state.l2_block_number, TEST_ANCHOR_BLOCK, "should return anchor block");
-        assert!(cache.is_some(), "cache should still be populated");
-    }
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_cold_start_uses_anchor_game_after_anchor_advance() {
-        let anchor_game = proxy_addr(0);
-        let anchor_block = TEST_BLOCK_INTERVAL;
-
-        let mut factory = MockDisputeGameFactory::with_games(vec![]);
-        factory.game_count_override = Some(1);
-        let pipeline = recovery_pipeline_full_with_anchor_game(
-            factory,
-            HashMap::new(),
-            anchor_block,
-            anchor_game,
-            TEST_BLOCK_INTERVAL,
-            TEST_BLOCK_INTERVAL,
-        );
-
-        let mut cache: Option<CachedRecovery> = None;
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        assert_eq!(state.parent_address, anchor_game, "advanced anchor game should be the parent");
-        assert_eq!(state.l2_block_number, anchor_block, "should propose after the live anchor");
-    }
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_reads_anchor_root_and_game_from_one_snapshot() {
-        let anchor_game = proxy_addr(0);
-        let anchor_root = B256::repeat_byte(0xAA);
-        let anchor_block = TEST_BLOCK_INTERVAL;
-        let cancel = CancellationToken::new();
-        let l1 = Arc::new(MockL1 { latest_block_number: TEST_L1_BLOCK_NUMBER });
-        let l2 = Arc::new(MockL2 { block_not_found: true, canonical_hash: None });
-        let rollup = Arc::new(MockRollupClient {
-            sync_status: test_sync_status(TEST_BLOCK_INTERVAL * 2, B256::ZERO),
-            output_roots: HashMap::new(),
-            max_safe_block: None,
-        });
-        let anchor_registry = Arc::new(SnapshotOnlyAnchorStateRegistry {
-            snapshot: base_proof_contracts::AnchorSnapshot {
-                anchor_root: base_proof_contracts::AnchorRoot {
-                    root: anchor_root,
-                    l2_block_number: anchor_block,
-                },
-                anchor_game,
-            },
-        });
-        let mut factory = MockDisputeGameFactory::with_games(vec![]);
-        factory.game_count_override = Some(1);
-
-        let pipeline = ProvingPipeline::new(
-            PipelineConfig {
-                submit_timeout: Some(std::time::Duration::from_secs(60)),
-                max_retries: 3,
-                recovery_scan_concurrency: 8,
-                tee_prover_registry_address: None,
-                driver: DriverConfig {
-                    block_interval: TEST_BLOCK_INTERVAL,
-                    intermediate_block_interval: TEST_BLOCK_INTERVAL,
-                    ..Default::default()
-                },
-            },
-            Arc::new(MockProofRequester::default()),
-            l1,
-            l2,
-            rollup,
-            anchor_registry,
-            Arc::new(factory),
-            Arc::new(MockAggregateVerifier::default()),
-            Arc::new(MockOutputProposer),
-            cancel,
-        );
-
-        let mut cache: Option<CachedRecovery> = None;
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        assert_eq!(state.parent_address, anchor_game);
-        assert_eq!(state.output_root, anchor_root);
-        assert_eq!(state.l2_block_number, anchor_block);
-    }
-
-    // ---- Recovery: forward walk ----
-
-    #[rstest]
-    #[case::single_game(1, 0, TEST_BLOCK_INTERVAL, "single game at first interval")]
-    #[case::chain_of_two(2, 1, TEST_BLOCK_INTERVAL * 2, "chain of two sequential games")]
-    #[case::chain_of_five(5, 4, TEST_BLOCK_INTERVAL * 5, "chain of five sequential games")]
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_forward_walk_chain(
-        #[case] game_count: usize,
-        #[case] expected_proxy_index: u64,
-        #[case] expected_block: u64,
-        #[case] scenario: &str,
-    ) {
-        let (factory, output_roots) = game_chain(game_count);
-        let pipeline = recovery_pipeline(factory, output_roots);
-
-        let mut cache: Option<CachedRecovery> = None;
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        assert_eq!(state.parent_address, proxy_addr(expected_proxy_index), "{scenario}");
-        assert_eq!(state.l2_block_number, expected_block, "{scenario}");
-        assert!(cache.is_some(), "{scenario}: cache should be populated");
-    }
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_forward_walk_stops_at_gap() {
-        // Game at block 512 exists, but no game at block 1024.
-        // Walk should stop after the first game.
-        let root_1 = B256::repeat_byte(0x01);
-        let extra_data_1 = encode_extra_data(TEST_BLOCK_INTERVAL, Address::ZERO, &[root_1]);
-
-        let mut factory = MockDisputeGameFactory::with_games(vec![]);
-        factory.game_count_override = Some(1);
-        factory.uuid_games.insert((TEST_GAME_TYPE, root_1, extra_data_1), proxy_addr(0));
-
-        let output_roots = HashMap::from([(TEST_BLOCK_INTERVAL, root_1)]);
-
-        let pipeline = recovery_pipeline(factory, output_roots);
-
-        let mut cache: Option<CachedRecovery> = None;
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        assert_eq!(state.parent_address, proxy_addr(0), "should stop at first game before gap");
-        assert_eq!(state.l2_block_number, TEST_BLOCK_INTERVAL);
-        assert_eq!(state.output_root, root_1);
-    }
-
-    // ---- Recovery: error propagation ----
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_propagates_games_lookup_failure() {
-        // A chain of 2 games exists, but factory.games() always fails.
-        // The walk should propagate the error as ProposerError::Contract.
-        let (mut factory, output_roots) = game_chain(2);
-        factory.games_should_fail = true;
-
-        let pipeline = recovery_pipeline(factory, output_roots);
-
-        let mut cache: Option<CachedRecovery> = None;
-        let result = pipeline.recover_latest_state(&mut cache).await;
-
-        assert!(result.is_err(), "games() failure should propagate");
-        let err = result.unwrap_err();
-        assert!(
-            matches!(err, ProposerError::Contract(_)),
-            "expected ProposerError::Contract, got {err:?}"
-        );
-    }
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_forward_walk_stops_at_safe_head() {
-        // 3 games exist on-chain, but the rollup node only has blocks up to
-        // block 2 * TEST_BLOCK_INTERVAL. The walk should verify games 0 and 1,
-        // then terminate gracefully when it can't fetch the output root for
-        // game 2's block (ahead of safe head).
-        let (factory, output_roots) = game_chain(3);
-
-        let cancel = CancellationToken::new();
-        let l1 = Arc::new(MockL1 { latest_block_number: TEST_L1_BLOCK_NUMBER });
-        let l2 = Arc::new(MockL2 { block_not_found: true, canonical_hash: None });
-        let rollup = Arc::new(MockRollupClient {
-            sync_status: test_sync_status(0, B256::ZERO),
-            output_roots,
-            max_safe_block: Some(TEST_BLOCK_INTERVAL * 2),
-        });
-        let anchor_registry = Arc::new(MockAnchorStateRegistry {
-            anchor_root: test_anchor_root(TEST_ANCHOR_BLOCK),
-            anchor_game: Address::ZERO,
-        });
-
-        let pipeline = ProvingPipeline::new(
-            PipelineConfig {
-                submit_timeout: Some(std::time::Duration::from_secs(60)),
-                max_retries: 1,
-                recovery_scan_concurrency: 8,
-                tee_prover_registry_address: None,
-                driver: DriverConfig {
-                    game_type: TEST_GAME_TYPE,
-                    block_interval: TEST_BLOCK_INTERVAL,
-                    intermediate_block_interval: TEST_BLOCK_INTERVAL,
-                    ..Default::default()
-                },
-            },
-            Arc::new(MockProofRequester::default()),
-            l1,
-            l2,
-            rollup,
-            anchor_registry,
-            Arc::new(factory),
-            Arc::new(MockAggregateVerifier::default()),
-            Arc::new(MockOutputProposer),
-            cancel,
-        );
-
-        let mut cache: Option<CachedRecovery> = None;
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        // Should stop after game 1 (block 1024), not reach game 2 (block 1536).
-        assert_eq!(state.parent_address, proxy_addr(1), "should stop at game 1");
-        assert_eq!(state.l2_block_number, TEST_BLOCK_INTERVAL * 2);
-    }
-
-    // ---- Recovery: caching ----
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_cache_hit_equal_game_count() {
-        let (factory, output_roots) = game_chain(1);
-        let game_proxy = proxy_addr(0);
-
-        let pipeline = recovery_pipeline(factory, output_roots);
-
-        // First call: cold start, populates the cache.
-        let mut cache: Option<CachedRecovery> = None;
-        let state1 = pipeline.recover_latest_state(&mut cache).await.unwrap();
-        assert!(cache.is_some(), "cache should be populated after first call");
-        assert_eq!(state1.parent_address, game_proxy);
-        assert_eq!(state1.l2_block_number, TEST_BLOCK_INTERVAL);
-        assert_eq!(cache.as_ref().unwrap().game_count, 1);
-
-        // Second call: same game_count → cached state returned without re-walk.
-        let state2 = pipeline.recover_latest_state(&mut cache).await.unwrap();
-        assert_eq!(state2.parent_address, state1.parent_address);
-        assert_eq!(state2.l2_block_number, state1.l2_block_number);
-        assert_eq!(state2.output_root, state1.output_root);
-    }
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_cache_incremental_on_count_increase() {
-        // Seed cache with game_count=1, state at game 0. Factory now has 2
-        // games. Anchor is still at block 0 (behind the cached tip at
-        // TEST_BLOCK_INTERVAL), so the walk resumes from the cached tip
-        // and only needs to discover game 1.
-        let (factory, output_roots) = game_chain(2);
-
-        let mut cache = Some(CachedRecovery {
-            game_count: 1,
-            state: RecoveredState {
-                parent_address: proxy_addr(0),
-                output_root: B256::repeat_byte(0x01),
-                l2_block_number: TEST_BLOCK_INTERVAL,
-            },
-        });
-
-        let pipeline = recovery_pipeline(factory, output_roots);
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        assert_eq!(state.parent_address, proxy_addr(1), "should find game 1 incrementally");
-        assert_eq!(state.l2_block_number, TEST_BLOCK_INTERVAL * 2);
-        assert_eq!(cache.as_ref().unwrap().game_count, 2, "cache should reflect new count");
-    }
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_cache_incremental_resumes_mid_chain() {
-        // Build a chain of 5 games. Seed cache at game 2 (game_count=3).
-        // Factory now has 5 games. The walk should resume from game 2's
-        // tip and discover games 3 and 4 without re-walking games 0–2.
-        let (factory, output_roots) = game_chain(5);
-
-        let mut cache = Some(CachedRecovery {
-            game_count: 3,
-            state: RecoveredState {
-                parent_address: proxy_addr(2),
-                output_root: B256::repeat_byte(0x03),
-                l2_block_number: TEST_BLOCK_INTERVAL * 3,
-            },
-        });
-
-        let pipeline = recovery_pipeline(factory, output_roots);
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        assert_eq!(state.parent_address, proxy_addr(4), "should reach game 4 from cached tip");
-        assert_eq!(state.l2_block_number, TEST_BLOCK_INTERVAL * 5);
-        assert_eq!(cache.as_ref().unwrap().game_count, 5);
-    }
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_cache_incremental_unrelated_games() {
-        // game_count increased (1 → 2) but the new game is not in our
-        // chain (no UUID entry at the next expected block). The incremental
-        // walk resumes from the cached tip, finds nothing, and returns the
-        // same state. This happens when another proposer creates a game
-        // with different parameters.
-        let (factory, output_roots) = game_chain(1);
-        // factory has game_count=1, but we'll seed cache as game_count=0
-        // so the code sees an increase (0 → 1). The walk from the anchor
-        // will find game 0. But to test the "unrelated game" path, we need
-        // game_count > cached_count and no new UUID at the next block.
-        //
-        // Seed cache at game 0, pretend game_count was 1. Factory reports
-        // game_count=2 (simulating someone else's unrelated game), but
-        // there's no UUID entry at block 2*TEST_BLOCK_INTERVAL.
-        let mut factory_with_extra_count = factory;
-        factory_with_extra_count.game_count_override = Some(2);
-
-        let pipeline = recovery_pipeline(factory_with_extra_count, output_roots);
-
-        let mut cache = Some(CachedRecovery {
-            game_count: 1,
-            state: RecoveredState {
-                parent_address: proxy_addr(0),
-                output_root: B256::repeat_byte(0x01),
-                l2_block_number: TEST_BLOCK_INTERVAL,
-            },
-        });
-
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        // Walk resumed from game 0, found no game at the next block,
-        // returned the same state.
-        assert_eq!(state.parent_address, proxy_addr(0), "should remain at game 0");
-        assert_eq!(state.l2_block_number, TEST_BLOCK_INTERVAL);
-        assert_eq!(cache.as_ref().unwrap().game_count, 2, "cache updated to new count");
-    }
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_cache_invalidated_by_count_decrease() {
-        // Seed cache with game_count=5. Factory now has only 1 game (reorg).
-        let (factory, output_roots) = game_chain(1);
-
-        let mut cache = Some(CachedRecovery {
-            game_count: 5,
-            state: RecoveredState {
-                parent_address: proxy_addr(99),
-                output_root: B256::repeat_byte(0xDD),
-                l2_block_number: 5 * TEST_BLOCK_INTERVAL,
-            },
-        });
-
-        let pipeline = recovery_pipeline(factory, output_roots);
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        assert_eq!(state.parent_address, proxy_addr(0), "reorg: should find the 1 remaining game");
-        assert_eq!(state.l2_block_number, TEST_BLOCK_INTERVAL);
-        assert_eq!(cache.as_ref().unwrap().game_count, 1, "reorg: cache should reflect new count");
-    }
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_cache_full_walk_when_anchor_past_tip() {
-        // Anchor is at block 2048 (past the cached tip at block 512).
-        // This simulates a governance intervention that advanced the
-        // anchor past the cached tip. A full walk from the new anchor
-        // is required.
-        let anchor_block = TEST_BLOCK_INTERVAL * 4; // block 2048
-        let (factory, output_roots) =
-            game_chain_full(1, anchor_block, TEST_BLOCK_INTERVAL, TEST_BLOCK_INTERVAL);
-
-        let mut cache = Some(CachedRecovery {
-            game_count: 0,
-            state: RecoveredState {
-                parent_address: proxy_addr(99), // stale — will be recomputed
-                output_root: B256::repeat_byte(0xDD),
-                l2_block_number: TEST_BLOCK_INTERVAL, // tip at 512, anchor at 2048
-            },
-        });
-
-        let pipeline = recovery_pipeline_full(
-            factory,
-            output_roots,
-            anchor_block,
-            TEST_BLOCK_INTERVAL,
-            TEST_BLOCK_INTERVAL,
-        );
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        // Anchor past cached tip → full walk from new anchor.
-        assert_eq!(state.parent_address, proxy_addr(0));
-        assert_eq!(state.l2_block_number, anchor_block + TEST_BLOCK_INTERVAL);
-    }
-
-    // ---- Recovery: intermediate roots with multiple checkpoints ----
-
-    /// Block intervals for recovery tests with multiple intermediate roots.
-    const RECOVERY_BI: u64 = 4;
-    const RECOVERY_IBI: u64 = 2;
-
-    #[tokio::test(flavor = "current_thread", start_paused = true)]
-    async fn test_recovery_forward_walk_with_intermediate_roots() {
-        // block_interval = 4, intermediate_block_interval = 2
-        // → intermediate_count = 2 (roots at parent+2 and parent+4)
-        //
-        // Two games: block 4 (parent = anchor) and block 8 (parent = game 0).
-        // Both have correct UUID including intermediate roots. Walk should
-        // traverse both games.
-        let (factory, output_roots) =
-            game_chain_full(2, TEST_ANCHOR_BLOCK, RECOVERY_BI, RECOVERY_IBI);
-
-        let pipeline = recovery_pipeline_full(
-            factory,
-            output_roots,
-            TEST_ANCHOR_BLOCK,
-            RECOVERY_BI,
-            RECOVERY_IBI,
-        );
-
-        let mut cache: Option<CachedRecovery> = None;
-        let state = pipeline.recover_latest_state(&mut cache).await.unwrap();
-
-        // Both games verified, walk should reach game 1.
-        assert_eq!(state.parent_address, proxy_addr(1));
-        assert_eq!(state.l2_block_number, RECOVERY_BI * 2);
     }
 
     // ---- Intermediate output root validation (submission) tests ----
