+++ crates/proof/proposer/src/proof_submitter.rs
@@ -371,9 +371,8 @@ where
     /// Returns intermediate block numbers between `starting_block_number` and
     /// the next proposal target, stepping by `intermediate_block_interval`.
     ///
-    /// Shared with [`crate::pipeline::ProvingPipeline`]'s recovery forward walk,
-    /// which calls through this method on the pipeline's owned submitter so the
-    /// stepping logic lives in exactly one place.
+    /// Used by submit validation to match the same checkpoint layout as
+    /// proposer recovery.
     pub fn intermediate_block_numbers(
         &self,
         starting_block_number: u64,
